Create Three-step LTA - Step2.1: Running latent class analysis for time 1 (baseline)
Next, we run LCAs at each timepoint, here you see the code for baseline analysis. The aim is to save the participant’s most probable latent class at each timepoint, while also recording the measurement errors for each latent class variable as recorded in the output section called “Logits for the Classification Probabilities for the Most likely latent class membership”. To do this the measurement models for each of the latent classes are fixed using the starting values from step 1.